Move cli to snage
