--- packages/cli/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "changelog_gen",
-  "version": "1.0.0",
-  "description": "Script designed to generate change log templates for the specified audiences",
-  "main": "gen_changelog.js",
-  "dependencies": {
-    "gray-matter": "^4.0.2",
-    "typescript": "^3.8.3",
-    "yargs": "^14.0.0",
-    "inquirer": "^7.1.0",
-    "moment": "^2.24.0",
-    "ts-node": "^8.8.2",
-    "fp-ts": "^2.5.3"
-  },
-  "devDependencies": {
-    "@types/inquirer": "^6.5.0",
-    "@types/yargs": "^15.0.4"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://git.omikron.net/scm/ff/fact-finder-framework.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
